pyproject.toml
[project]
name = "autotransgal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "manga-ocr>=0.1.14",
    "mss>=10.1.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "pynput>=1.8.1",
    "rapidocr-onnxruntime>=1.4.4",
]

[project.scripts]
cli = "autotransgal.cli:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autotransgal"]
